Made updates as per WestDiscGolf comments on pull request
