+++ Fitbit.Portable/IFitbitClient.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Fitbit.Api.Portable.Models;
 using Fitbit.Models;
 
 namespace Fitbit.Api.Portable
@@ -34,18 +35,18 @@ namespace Fitbit.Api.Portable
         Task<ApiSubscription> AddSubscriptionAsync(APICollectionType apiCollectionType, string uniqueSubscriptionId, string subscriberId = default(string));
         Task DeleteSubscriptionAsync(APICollectionType collection, string uniqueSubscriptionId, string subscriberId = null);
         Task<ActivityLog> LogActivityAsync(ActivityLog model);
-        Task<List<HeartActivitiesTimeSeries>> GetHeartRateTimeSeries(string date, string period,
+        Task<List<HeartActivitiesTimeSeries>> GetHeartRateTimeSeriesAsync(string date, string period,
             string encodedUserId = default(string));
-        Task<List<HeartActivitiesTimeSeries>> GetHeartRateTimeSeries(string baseDate, DateTime endDate,
+        Task<List<HeartActivitiesTimeSeries>> GetHeartRateTimeSeriesAsync(string baseDate, DateTime endDate,
             string encodedUserId = default(string));
-        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeries(string date, DateTime endDate, string detailLevel,
+        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeriesAsync(string date, DateTime endDate, string detailLevel,
             string encodedUserId = default(string));
-        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeries(string date, DateTime endDate, string detailLevel,
+        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeriesAsync(string date, DateTime endDate, string detailLevel,
             TimeSpan startTime, TimeSpan endTime, string encodedUserId = default(string));
-        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeries(string date, string detailLevel,
+        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeriesAsync(string date, string detailLevel,
             string encodedUserId = default(string));
-        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeries(string date, string detailLevel, TimeSpan startTime,
+        Task<HeartActivitiesIntraday> GetHeartRateIntradayTimeSeriesAsync(string date, string detailLevel, TimeSpan startTime,
             TimeSpan endTime, string encodedUserId = default(string));
-        Task<List<ActivityList>> GetActivityLogsList(DateTime? beforeDate, DateTime? afterDate, int limit = 20, string encodedUserId = default(string));
+        Task<List<ActivityList>> GetActivityLogsListAsync(DateTime? beforeDate, DateTime? afterDate, int limit = 20, string encodedUserId = default(string));
     }
 }
