+++ Fitbit.Portable/Models/Source.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace Fitbit.Api.Portable.Models
+{
+    public class Source
+    {
+        [JsonProperty(PropertyName = "id")]
+        public string Id { get; set; }
+
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
+
+        [JsonProperty(PropertyName = "type")]
+        public string Type { get; set; }
+
+        [JsonProperty(PropertyName = "url")]
+        public string Url { get; set; }
+    }
+}
